requirements.txt

streamlit
pytesseract
pdf2image
opencv-python-headless
python-docx
